backend/requirements.txt
flask==3.0.2
flask-cors==4.0.0
python-dotenv==1.0.1
PyGithub==2.2.0
google-generativeai==0.3.2
requests==2.31.0
pyjwt==2.8.0
werkzeug==3.0.1